Value of DH prime changed to RFC-5114

+++ app/src/main/java/com/example/juanperezdealgaba/sac/AsyncOpenSSL.java
@@ -160,7 +160,7 @@ public class AsyncOpenSSL extends AsyncTask<OpenSSLParamsTest,Void,String> {
             text.append("**********MD5************\n");
 
                 for (int i = 0; i < repetitions; i++) {
-                    double[] testMD5 = test.MD5();
+                    int[] testMD5 = test.MD5();
                     System.out.println("Time to generate hash:" + testMD5[1] + "ns\n");
                     writer.write("Time to generate hash:" + testMD5[1] + "ns\n");
                     text.append("Time to generate hash:" + testMD5[1] + "ns\n");
