+++ app/src/main/java/com/example/juanperezdealgaba/sac/OpenSSL.java
@@ -15,7 +15,7 @@ public class OpenSSL {
 
     public native int[] ECDH();
 
-    public native double[] MD5();
+    public native int[] MD5();
 
     public native int[] AESCTR();
